requirements.txt
langgraph
langchain
langchain-openai
langchain-community
fastapi
uvicorn
pandas
matplotlib
seaborn
numpy
python-dotenv
scikit-learn
pytest
pytest-cov
pytest-mock
streamlit
openpyxl
plotly
chardet
scipy

